+++ python/mangadap/proc/bandpassfilter.py
@@ -393,7 +393,7 @@ def passband_integral(x, y, passband=None, borders=False, log=False, base=10.0, 
                     else numpy.sum(_y*inband_dx, axis=0)
 
     # Reshape the returned array if only one passband provided
-    return integral if _passband.shape[0] > 1 else integral[0]
+    return integral if numpy.atleast_1d(passband).ndim == 2 else integral[0]
 
 
 def passband_integrated_width(x, y, passband=None, borders=False, log=False, base=10.0):
@@ -544,7 +544,7 @@ def passband_weighted_sdev(x, y, z, passband=None, borders=False, yerr=None, zer
         return sigma, None
 
     # Construct error
-    sigma_error = numpy.zeros_like(sigma, dtype=float)
+    sigma_error = numpy.zeros_like(sigma, dtype=float) if isinstance(sigma, numpy.ndarray) else 0.
 
     # Get the integrand for the y errors
     if yerr is not None:
@@ -560,7 +560,7 @@ def passband_weighted_sdev(x, y, z, passband=None, borders=False, yerr=None, zer
     if zerr is not None:
         if zerr.shape != z.shape:
             raise ValueError('Incorrect shape for z errors.')
-        ze_integrand = 2*(z[:,None]-mu[None,:])*(y*zerr)[:,None]
+        ze_integrand = 2*(z[:,None]-numpy.atleast_1d(mu)[None,:])*(y*zerr)[:,None]
         sigma_error += numpy.square(passband_integral(x, ze_integrand, passband=passband,
                                                       borders=borders, log=log, base=base,
                                                       quad=True))